refactor(app): move useProtocolAnalysisErrors
