+++ app/src/organisms/Devices/hooks/index.ts
@@ -9,7 +9,6 @@ export * from './useLPCSuccessToast'
 export * from './usePipetteOffsetCalibrations'
 export * from './usePipetteOffsetCalibration'
 export * from './useTipLengthCalibrations'
-export * from './useProtocolAnalysisErrors'
 export * from './useRunStartedOrLegacySessionInProgress'
 export * from './useTrackCreateProtocolRunEvent'
 export * from './useRunStatuses'
