+++ app/src/pages/ODD/ProtocolSetup/__tests__/ProtocolSetup.test.tsx
@@ -25,7 +25,6 @@ import {
 import { i18n } from '/app/i18n'
 import { useToaster } from '/app/organisms/ToasterOven'
 import { mockRobotSideAnalysis } from '/app/molecules/Command/__fixtures__'
-import { useProtocolAnalysisErrors } from '/app/organisms/Devices/hooks'
 import { useAttachedModules } from '/app/resources/modules'
 import { useRobotType } from '/app/redux-resources/robots'
 import { useTrackProtocolRunEvent } from '/app/redux-resources/analytics'
@@ -62,6 +61,7 @@ import {
   useRunCreatedAtTimestamp,
   useLPCDisabledReason,
   useModuleCalibrationStatus,
+  useProtocolAnalysisErrors,
 } from '/app/resources/runs'
 import { mockConnectableRobot } from '/app/redux/discovery/__fixtures__'
 import { mockRunTimeParameterData } from '/app/organisms/ODD/ProtocolSetup/__fixtures__'
@@ -102,7 +102,6 @@ vi.mock('react-router-dom', async importOriginal => {
 
 vi.mock('@opentrons/react-api-client')
 vi.mock('/app/organisms/LabwarePositionCheck/useLaunchLPC')
-vi.mock('/app/organisms/Devices/hooks')
 vi.mock('/app/organisms/ODD/ProtocolSetup', async importOriginal => {
   const ACTUALS = ['ProtocolSetupStep']
   const actual = await importOriginal<object>()
