+++ app/src/redux-resources/analytics/hooks/__tests__/useRobotAnalyticsData.test.tsx
@@ -44,7 +44,7 @@ const ROBOT_SERIAL_NUMBER = 'OT123'
 let wrapper: React.FunctionComponent<{ children: React.ReactNode }>
 let store: Store<any> = createStore(vi.fn(), {})
 
-describe('useProtocolAnalysisErrors hook', () => {
+describe('useRobotAnalyticsData hook', () => {
   beforeEach(() => {
     store = createStore(vi.fn(), {})
     store.dispatch = vi.fn()
