+++ app/src/resources/runs/__tests__/useProtocolAnalysisErrors.test.tsx
@@ -7,10 +7,10 @@ import {
   useProtocolQuery,
 } from '@opentrons/react-api-client'
 
-import { useProtocolAnalysisErrors } from '..'
-import { useNotifyRunQuery } from '/app/resources/runs'
+import { useProtocolAnalysisErrors } from '../useProtocolAnalysisErrors'
+import { useNotifyRunQuery } from '../useNotifyRunQuery'
 
-import { RUN_ID_2 } from '/app/resources/runs/__fixtures__'
+import { RUN_ID_2 } from '../__fixtures__'
 
 import type { UseQueryResult } from 'react-query'
 import type { Run, Protocol } from '@opentrons/api-client'
@@ -20,7 +20,7 @@ import type {
 } from '@opentrons/shared-data'
 
 vi.mock('@opentrons/react-api-client')
-vi.mock('/app/resources/runs')
+vi.mock('../useNotifyRunQuery')
 
 describe('useProtocolAnalysisErrors hook', () => {
   beforeEach(() => {
