+++ app/src/resources/runs/index.ts
@@ -26,3 +26,4 @@ export * from './useLPCDisabledReason'
 export * from './useUnmatchedModulesForProtocol'
 export * from './useModuleRenderInfoForProtocolById'
 export * from './useModuleCalibrationStatus'
+export * from './useProtocolAnalysisErrors'
